Purge button module and some dropdowns styling
