+++ i18n/en.json
@@ -44,6 +44,8 @@
   "tgui-site-nav-label": "Site",
   "tgui-prefs-talkpage": "[[mw:Talk:Reading/Web/Desktop_Improvements|Discussion]]",
   "tooltip-tgui-anon-user-menu-title": "More options",
+  "tgui-purge-button-label": "Purge Cache",
+  "tgui-purge-button-title": "Purge this page cache [F4]",
 
   "tgui-theme-name": "Color",
   "tgui-theme-description": "Change the theme of the wiki",
