+++ resources/skins.tgui.preferences/skins.tgui.preferences.less
@@ -53,7 +53,7 @@
     overflow: hidden;
     opacity: 0;
     right: 0;
-    min-width: 200px;
+    min-width: 175px;
     background-color: var(--color-bg);
     color: var(--color-text);
     border: var(--border-classic);
@@ -149,8 +149,9 @@
       display: flex;
       align-items: center;
       justify-content: center;
-      font-weight: var(--font-weight-medium);
       height: 32px;
+      padding: 0 0.5em;
+      font-weight: var(--font-weight-medium);
       background-color: var(--color-primary);
       color: var(--color-white);
       border-radius: var(--border-radius-small);
