+++ resources/skins.tgui.styles/components/MenuDropdownCheckbox.less
@@ -67,7 +67,7 @@
     visibility: visible;
     opacity: 1;
     height: auto;
-    transform: translate(0, 1em);
+    transform: translate(0, 0.75em);
   }
 
   :not(:checked) > & {
